Fixed broken buttons.
